refactor(auto_source): colocate discovery modules with scrapers
