+++ lib/html2rss/auto_source/scraper/discovery.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Html2rss
+  class AutoSource
+    module Scraper
+      ##
+      # Card/list discovery helpers used by AutoSource scrapers before field extraction.
+      #
+      # HtmlExtractor fills article fields from a container; discovery finds those containers.
+      module Discovery
+      end
+    end
+  end
+end
