+++ spec/lib/html2rss/auto_source/scraper/discovery/class_clustering_spec.rb
@@ -2,7 +2,7 @@
 
 require 'nokogiri'
 
-RSpec.describe Html2rss::AutoSource::Scraper::Html::ClassClustering do
+RSpec.describe Html2rss::AutoSource::Scraper::Discovery::ClassClustering do
   describe '.call' do
     let(:html) do
       <<~HTML
